Support for semicolon comment delimiter
INI style configuration files should support the semicolon ';' character, when
it is the first character on the line only, as a comment delimiter. Because
the '#' style comments are not allowed in strict MS INI files, this gives an
opening for files to be compatible in both and still have comments.